Drop OrderedDict, use built-in dict
Some classes in _pyahocorasick relied on OrderedDict for as_dict which
used an OrderedDict purely for the __repr__, so make them use just
dict. Python 3.6 even has an implementation detail that the built-in
dict behaves like an OrderedDict, so all tests pass there.

Alternative is to bring OrderedDict from collections.py and
_collections_abc.py which still would not transpile with Transcrypt
because it uses decorators heavily.

Signed-off-by: Aleksandr Lisianoi <all3fox@gmail.com>