Drop custom transcrypt_true() function, use -t switch instead
Turns out the transcrypt_true function returned true for empty
dictionaries (when false is expected) and that is because Transcrypt
wraps dictionaries into "true" dictionaries in order to distinguis
them from plain javascript classes. So, instead of crafting custom
comparison logic, use the -t switch which bridges this runtime gap.

Signed-off-by: Aleksandr Lisianoi <all3fox@gmail.com>