Licensing.parse: Raise ExpressionParseError instead of ParseError
ExpressionParseError is a subclass of both ParseError and
ExpressionError. This allows the user to do `except ParseError`
as they had done before, and nothing will break. If the user does
`except ExpressionError`, the ParseErrors will now also be caught.

The advantage of doing this is that `except ExpressionError` now
catches all exceptions coming out of this method.

This is therefore a breaking change.

Signed-off-by: Carmen Bianca Bakker <carmen@carmenbianca.eu>